MA0076 detect more cases
