+++ src/Meziantou.Analyzer/Rules/DoNotUseImplicitCultureSensitiveToStringAnalyzer.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Immutable;
 using Meziantou.Analyzer.Configurations;
+using Meziantou.Analyzer.Internals;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Operations;
@@ -46,204 +47,123 @@ public sealed class DoNotUseImplicitCultureSensitiveToStringAnalyzer : Diagnosti
         context.EnableConcurrentExecution();
         context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.None);
 
-        context.RegisterOperationAction(AnalyzeBinaryOperation, OperationKind.Binary);
-        context.RegisterOperationAction(AnalyzeInterpolatedString, OperationKind.InterpolatedString);
-        context.RegisterOperationAction(AnalyzeInvocation, OperationKind.Invocation);
+        context.RegisterCompilationStartAction(context =>
+        {
+            var analyzerContext = new AnalyzerContext(context.Compilation);
+
+            context.RegisterOperationAction(analyzerContext.AnalyzeBinaryOperation, OperationKind.Binary);
+            context.RegisterOperationAction(analyzerContext.AnalyzeInterpolatedString, OperationKind.InterpolatedString);
+            context.RegisterOperationAction(AnalyzerContext.AnalyzeInvocation, OperationKind.Invocation);
+        });
     }
 
-    private static void AnalyzeInvocation(OperationAnalysisContext context)
+    private sealed class AnalyzerContext
     {
-        var operation = (IInvocationOperation)context.Operation;
-        if (IsExcludedMethod(context, s_objectToStringRule, operation))
-            return;
+        private readonly CultureSensitiveFormattingContext _cultureSensitiveContext;
 
-        if (operation.TargetMethod.Name == "ToString" && operation.TargetMethod.ContainingType.IsObject() && operation.TargetMethod.Parameters.Length == 0)
+        public AnalyzerContext(Compilation compilation)
         {
-            if (operation.Instance != null && operation.Instance.Type.IsObject())
-            {
-                context.ReportDiagnostic(s_objectToStringRule, operation);
-            }
-        }
-    }
-
-    private static void AnalyzeBinaryOperation(OperationAnalysisContext context)
-    {
-        var operation = (IBinaryOperation)context.Operation;
-        if (operation.OperatorKind != BinaryOperatorKind.Add)
-            return;
-
-        if (!operation.Type.IsString())
-            return;
-
-        if (operation.ConstantValue.HasValue)
-            return;
-
-        if (IsExcludedMethod(context, s_stringConcatRule, operation))
-            return;
-
-        if (!IsValidOperand(operation.LeftOperand))
-        {
-            context.ReportDiagnostic(s_stringConcatRule, operation.LeftOperand);
+            _cultureSensitiveContext = new CultureSensitiveFormattingContext(compilation);
         }
 
-        if (!IsValidOperand(operation.RightOperand))
+        public static void AnalyzeInvocation(OperationAnalysisContext context)
         {
-            context.ReportDiagnostic(s_stringConcatRule, operation.RightOperand);
-        }
-    }
-
-    private static void AnalyzeInterpolatedString(OperationAnalysisContext context)
-    {
-        // Check if parent is InterpolatedString.Invariant($"") or conversion to string?
-        var operation = (IInterpolatedStringOperation)context.Operation;
-
-        if (operation.ConstantValue.HasValue)
-            return;
-
-        if (IsExcludedMethod(context, s_stringInterpolationRule, operation))
-            return;
-
-        var parent = operation.Parent;
-        if (parent is IConversionOperation conversionOperation)
-        {
-            // `FormattableString _ = $""` is valid whereas `string _ = $""` may not be
-            if (conversionOperation.Type.IsEqualTo(context.Compilation.GetBestTypeByMetadataName("System.FormattableString")))
+            var operation = (IInvocationOperation)context.Operation;
+            if (IsExcludedMethod(context, s_objectToStringRule, operation))
                 return;
-        }
 
-        foreach (var part in operation.Parts.OfType<IInterpolationOperation>())
-        {
-            var expression = part.Expression;
-            var type = expression.Type;
-            if (expression == null || type == null)
-                continue;
-
-            if (IsFormattableType(context.Compilation, type) && !IsConstantPositiveNumber(expression))
+            if (operation.TargetMethod.Name == "ToString" && operation.TargetMethod.ContainingType.IsObject() && operation.TargetMethod.Parameters.Length == 0)
             {
-                context.ReportDiagnostic(s_stringInterpolationRule, part);
+                if (operation.Instance != null && operation.Instance.Type.IsObject())
+                {
+                    context.ReportDiagnostic(s_objectToStringRule, operation);
+                }
             }
         }
-    }
 
-    private static bool IsExcludedMethod(OperationAnalysisContext context, DiagnosticDescriptor descriptor, IOperation operation)
-    {
-        // ToString show culture-sensitive data by default
-        if (operation?.GetContainingMethod(context.CancellationToken)?.Name == "ToString")
+        public void AnalyzeBinaryOperation(OperationAnalysisContext context)
         {
-            return context.Options.GetConfigurationValue(operation.Syntax.SyntaxTree, descriptor.Id + ".exclude_tostring_methods", defaultValue: true);
-        }
+            var operation = (IBinaryOperation)context.Operation;
+            if (operation.OperatorKind != BinaryOperatorKind.Add)
+                return;
 
-        return false;
-    }
+            if (!operation.Type.IsString())
+                return;
 
-    private static bool IsValidOperand(IOperation operand)
-    {
-        // Implicit conversion from a type number
-        if (operand is null)
-            return true;
+            if (operation.ConstantValue.HasValue)
+                return;
 
-        if (operand is IConversionOperation conversion && conversion.IsImplicit && conversion.Type.IsObject() && conversion.Operand.Type != null)
-        {
-            if (IsFormattableType(operand.SemanticModel!.Compilation, conversion.Operand.Type) && !IsConstantPositiveNumber(conversion.Operand))
-                return false;
-        }
+            if (IsExcludedMethod(context, s_stringConcatRule, operation))
+                return;
 
-        return true;
-    }
-
-    private static bool IsFormattableType(Compilation compilation, ITypeSymbol typeSymbol)
-    {
-        var iformattableSymbol = compilation.GetBestTypeByMetadataName("System.IFormattable");
-        if (typeSymbol.Implements(iformattableSymbol))
-        {
-            if (typeSymbol.IsEnumeration())
-                return false;
-
-            if (typeSymbol.SpecialType == SpecialType.System_Byte)
-                return false;
-
-            if (typeSymbol.SpecialType == SpecialType.System_Char)
-                return false;
-
-            if (typeSymbol.SpecialType == SpecialType.System_UInt16)
-                return false;
-
-            if (typeSymbol.SpecialType == SpecialType.System_UInt32)
-                return false;
-
-            if (typeSymbol.SpecialType == SpecialType.System_UInt64)
-                return false;
-
-            if (typeSymbol.IsEqualTo(compilation.GetBestTypeByMetadataName("System.UInt128")))
-                return false;
-            
-            if (typeSymbol.IsEqualTo(compilation.GetBestTypeByMetadataName("System.UIntPtr")))
-                return false;
-            
-            if (typeSymbol.IsEqualTo(compilation.GetBestTypeByMetadataName("System.TimeSpan")))
-                return false;
-
-            if (typeSymbol.IsEqualTo(compilation.GetBestTypeByMetadataName("System.Guid")))
-                return false;
-
-            if (typeSymbol.IsOrInheritFrom(compilation.GetBestTypeByMetadataName("System.Version")))
-                return false;
-
-            if (typeSymbol.IsEqualTo(compilation.GetBestTypeByMetadataName("System.Windows.FontStretch")))
-                return false;
-
-            if (typeSymbol.IsOrInheritFrom(compilation.GetBestTypeByMetadataName("System.Windows.Media.Brush")))
-                return false;
-
-            return true;
-        }
-
-        return false;
-    }
-
-    // Only negative numbers are culture-sensitive (negative sign)
-    // For instance, https://source.dot.net/#System.Private.CoreLib/Int32.cs,8d6f2d8bc0589463
-    private static bool IsConstantPositiveNumber(IOperation operation)
-    {
-        if (operation.Type != null && operation.ConstantValue.HasValue)
-        {
-            var constantValue = operation.ConstantValue.Value;
-            bool? result = operation.Type.SpecialType switch
+            if (!IsNonCultureSensitiveOperand(operation.LeftOperand))
             {
-                SpecialType.System_Byte => true,
-                SpecialType.System_SByte => (sbyte)constantValue! >= 0,
-                SpecialType.System_Int16 => (short)constantValue! >= 0,
-                SpecialType.System_Int32 => (int)constantValue! >= 0,
-                SpecialType.System_Int64 => (long)constantValue! >= 0,
-                SpecialType.System_UInt16 => true,
-                SpecialType.System_UInt32 => true,
-                SpecialType.System_UInt64 => true,
-                _ => null,
-            };
-            if (result.HasValue)
-                return result.Value;
+                context.ReportDiagnostic(s_stringConcatRule, operation.LeftOperand);
+            }
+
+            if (!IsNonCultureSensitiveOperand(operation.RightOperand))
+            {
+                context.ReportDiagnostic(s_stringConcatRule, operation.RightOperand);
+            }
         }
 
-        if (operation is IMemberReferenceOperation memberReferenceOperation)
+        public void AnalyzeInterpolatedString(OperationAnalysisContext context)
         {
-            if (memberReferenceOperation.Member.Name == "Count")
-                return true;
+            // Check if parent is InterpolatedString.Invariant($"") or conversion to string?
+            var operation = (IInterpolatedStringOperation)context.Operation;
 
-            if (memberReferenceOperation.Member.Name == "Length")
-                return true;
+            if (operation.ConstantValue.HasValue)
+                return;
 
-            if (memberReferenceOperation.Member.Name == "LongthLength")
-                return true;
+            if (IsExcludedMethod(context, s_stringInterpolationRule, operation))
+                return;
+
+            var parent = operation.Parent;
+            if (parent is IConversionOperation conversionOperation)
+            {
+                // `FormattableString _ = $""` is valid whereas `string _ = $""` may not be
+                if (conversionOperation.Type.IsEqualTo(context.Compilation.GetBestTypeByMetadataName("System.FormattableString")))
+                    return;
+            }
+
+            foreach (var part in operation.Parts.OfType<IInterpolationOperation>())
+            {
+                var expression = part.Expression;
+                var type = expression.Type;
+                if (expression == null || type == null)
+                    continue;
+
+                if (_cultureSensitiveContext.IsCultureSensitiveType(type, format: part.FormatString, instance: expression))
+                {
+                    context.ReportDiagnostic(s_stringInterpolationRule, part);
+                }
+            }
         }
-        else if (operation is IInvocationOperation invocationOperation)
+
+        private static bool IsExcludedMethod(OperationAnalysisContext context, DiagnosticDescriptor descriptor, IOperation operation)
         {
-            if (invocationOperation.TargetMethod.Name == "Count")
-                return true;
+            // ToString show culture-sensitive data by default
+            if (operation?.GetContainingMethod(context.CancellationToken)?.Name == "ToString")
+            {
+                return context.Options.GetConfigurationValue(operation.Syntax.SyntaxTree, descriptor.Id + ".exclude_tostring_methods", defaultValue: true);
+            }
 
-            if (invocationOperation.TargetMethod.Name == "LongCount")
-                return true;
+            return false;
         }
 
-        return false;
+        private bool IsNonCultureSensitiveOperand(IOperation operand)
+        {
+            // Implicit conversion from a type number
+            if (operand is null)
+                return true;
+
+            if (operand is IConversionOperation conversion && conversion.IsImplicit && conversion.Type.IsObject() && conversion.Operand.Type != null)
+            {
+                if (_cultureSensitiveContext.IsCultureSensitiveType(conversion.Operand.Type, format: null, instance: conversion.Operand))
+                    return false;
+            }
+
+            return true;
+        }
     }
 }
